equipment_knowledge_graph_project/requirements.txt
# Web框架
Django==4.2.7
FastAPI==0.104.1
uvicorn==0.24.0

# 数据库
neo4j==5.14.1
pymysql==1.1.0
sqlalchemy==2.0.23

# NLP处理
jieba==0.42.1
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1

# 深度学习
tensorflow==2.15.0
keras==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# 大模型相关
modelscope==1.9.5
sentence-transformers==2.2.2
langchain==0.0.350
openai==1.3.7

# 数据采集
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
PyPDF2==3.0.1
pdfplumber==0.10.2

# 数据处理
openpyxl==3.1.2
xlrd==2.0.1
python-docx==1.1.0

# 可视化
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
networkx==3.2.1
py2neo==2021.2.3

# API和Web
flask==3.0.0
flask-cors==4.0.0
aiohttp==3.9.1
httpx==0.25.2

# 工具库
tqdm==4.66.1
click==8.1.7
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# 部署相关
gunicorn==21.2.0
docker==6.1.3
kubernetes==28.1.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1